+++ package.json
@@ -17,6 +17,9 @@
     "metadata-lint.js",
     "scene-character-normalization.js",
     "review-bundles.js",
+    "review-bundles-planner.js",
+    "review-bundles-renderer.js",
+    "review-bundles-writer.js",
     "prose-styleguide.js",
     "prose-styleguide-drift.js",
     "prose-styleguide-skill.js",
